Refactoring blog schema and fixing pic function
